Move DJ to separate test run.  Better error capture.
