+++ .travis.yml
@@ -16,6 +16,7 @@ gemfile:
   - gemfiles/frameworks.gemfile
   - gemfiles/rails32.gemfile
   - gemfiles/rails42.gemfile
+  - gemfiles/delayed_job.gemfile
 
 matrix:
   exclude:
@@ -26,6 +27,10 @@ matrix:
       gemfile: gemfiles/rails41.gemfile
     - rvm: 1.8.7
       gemfile: gemfiles/rails42.gemfile
+
+    # Don't test DelayedJob under JRuby
+    - rvm: jruby-9.0.0.0
+      gemfile: gemfiles/delayed_job.gemfile
 
     # Other frameworks tested on Ruby 1.9.3 or higher
     - rvm: 1.8.7
