+++ test/servers/delayed_job.rb
@@ -1,10 +1,14 @@
 # Copyright (c) 2015 AppNeta, Inc.
 # All rights reserved.
 
-require "rails"
+require "rails/all"
 require "delayed_job"
+require "action_controller/railtie" # require more if needed
+require 'rack/handler/puma'
+require File.expand_path(File.dirname(__FILE__) + '/../models/widget')
 
 TraceView.logger.level = Logger::DEBUG
+TraceView.logger.info "[traceview/info] Starting background utility rails app on localhost:8140."
 
 if ENV.key?('TRAVIS_PSQL_PASS')
   DJ_DB_URL = "postgres://postgres:#{ENV['TRAVIS_PSQL_PASS']}@127.0.0.1:5432/travis_ci_test"
@@ -20,6 +24,59 @@ unless ActiveRecord::Base.connection.table_exists? :delayed_jobs
   ActiveRecord::Migration.run(CreateDelayedJobs)
 end
 
+unless ActiveRecord::Base.connection.table_exists? 'widgets'
+  ActiveRecord::Migration.run(CreateWidgets)
+end
+
+class Rails40MetalStack < Rails::Application
+  routes.append do
+    get "/hello/world" => "hello#world"
+    get "/hello/metal" => "ferro#world"
+  end
+
+  # Enable cache classes. Production style.
+  config.cache_classes = true
+  config.eager_load = false
+
+  # uncomment below to display errors
+  # config.consider_all_requests_local = true
+
+  config.active_support.deprecation = :stderr
+
+  # Here you could remove some middlewares, for example
+  # Rack::Lock, ActionDispatch::Flash and  ActionDispatch::BestStandardsSupport below.
+  # The remaining stack is printed on rackup (for fun!).
+  # Rails API has config.middleware.api_only! to get
+  # rid of browser related middleware.
+  config.middleware.delete "Rack::Lock"
+  config.middleware.delete "ActionDispatch::Flash"
+  config.middleware.delete "ActionDispatch::BestStandardsSupport"
+
+  # We need a secret token for session, cookies, etc.
+  config.secret_token = "49837489qkuweoiuoqwehisuakshdjksadhaisdy78o34y138974xyqp9rmye8yrpiokeuioqwzyoiuxftoyqiuxrhm3iou1hrzmjk"
+  config.secret_key_base = "2048671-96803948"
+end
+
+#################################################
+#  Controllers
+#################################################
+
+class HelloController < ActionController::Base
+  def world
+    render :text => "Hello world!"
+  end
+end
+
+class FerroController < ActionController::Metal
+  include AbstractController::Rendering
+
+  def world
+    render :text => "Hello world!"
+  end
+end
+
+Delayed::Job.delete_all
+
 @worker_options = {
   :min_priority => ENV['MIN_PRIORITY'],
   :max_priority => ENV['MAX_PRIORITY'],
@@ -32,10 +89,14 @@ end
 
 TraceView.logger.info "[traceview/servers] Starting up background DelayedJob."
 
+#Delayed::Worker.delay_jobs = false
+Delayed::Worker.max_attempts = 0
+Delayed::Worker.sleep_delay = 30
+
 Thread.new do
-  Delayed::Job.delete_all
   Delayed::Worker.new(@worker_options).start
 end
 
 # Allow it to boot
-sleep 2
+TraceView.logger.info "[traceview/servers] Waiting 5 seconds for DJ to boot..."
+sleep 5
